blueprinthub/tests/python_cli/pyproject.toml
[tool.poetry]
name = "mycli"
version = "0.1.0"
description = "A CLI tool built with BlueprintHub"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
typer = "^0.9.0"
pyyaml = "^6.0.1"
loguru = "^0.7.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"